Added first test flight files.
I did not close the logger upon detection of landing in the event that the FSM automatically detected landing early for some reason. Therefore the data is long for this first test flight. This flight was on a small motor and in a smaller rocket just to test the FSM.